Performance Improvement of create login in BBF (#507)
Extension PR : babelfish-for-postgresql/babelfish_extensions#3291

In Major version 16 Babelfish introduced a new role bbf_admin_role which is a member of all the roles which created implicitly from TSQL. This increases the Load on any Grant/Revoke operations that happen since in BBF to "select the role through which it gets permission to current user to grant", is inefficient and loops over members of members to find the all members of the role even though the required role is find.
This commit optimizes the selection of role with is_admin_option true for babelfish by escaping the search for any TSQL roles conditionally.

Issues Resolved
BABEL-5349

(cherry picked from commit 3b4b4eee4d25d4565d50235610121cfe8340954b)